Add adapter_config.json and optimizer.pt for various model checkpoints
This commit includes the addition of adapter_config.json and optimizer.pt files for several model checkpoints in the bert_lora_peft directory. These files are configurations that will be used in different checkpoints which have been created during the model training process. Please note that the burden of model size management is reduced since only the diffs will be stored and transmitted in Git.