Update model and training parameters in bert_lora_peft notebook
The commit updates the bert_lora_peft notebook, changing the training model from RoBERTa to bert-base-uncased. The use of Lora for efficiency is also removed. Further, parameters for training_args are updated and the push_to_hub functionality is deleted. Finally, the pretrained model is saved with a new name.