Add PEFT training and evaluation for Roberta model
This commit introduces a new Jupyter notebook where a Roberta model is trained using the Principle of Exponential Frequency Tuning (PEFT) strategy. The model is tested on the IMDb dataset. Performance metrics are computed and a random sample is used at the end of the notebook to demonstrate usage of the model.